asdf

+++ app/models/Log.scala
@@ -9,7 +9,7 @@ import play.api.db.DB
   * Created by pdeboer on 20/11/15.
   */
 object Log {
-	def createEntry(url: String, ip: String, userId: Long): Unit = {
+	def createEntry(url: String, ip: String, userId: Int): Unit = {
 		DB.withConnection { implicit c =>
 			SQL("INSERT INTO log (accesstime, url, ip, user) VALUES (NOW(), {url}, {ip}, {userId})").on(
 				'url -> url,
